requirements.txt
torch==2.0.1
torchaudio==2.0.2
numpy>=1.24.3
scipy>=1.10.0
gradio>=3.50.2
soundfile>=0.12.1
PyYAML>=6.0.1
unidecode>=1.3.6
python-dotenv>=1.0.0
colorama>=0.4.6
numba>=0.57.1
transformers>=4.30.2
einops>=0.6.1
rotary-embedding-torch>=0.2.3
mishkal
